+++ src/is_solvable.c
@@ -1,6 +1,6 @@
 #include "npuzzle.h"
 
-static void fill_numbers(t_env *env, int **puzzle, int *numbers)
+static void fill_numbers(t_env *env, unsigned short **puzzle, unsigned short *numbers)
 {
 	int y;
 	int x;
@@ -20,7 +20,7 @@ static void fill_numbers(t_env *env, int **puzzle, int *numbers)
 	}
 }
 
-static int get_inversions(t_env *env, int *numbers)
+static int get_inversions(t_env *env, unsigned short *numbers)
 {
 	int total;
 	int i;
@@ -47,7 +47,7 @@ static int get_inversions(t_env *env, int *numbers)
 	return (total);
 }
 
-static int blank_line(t_env *env, int *tab)
+static int blank_line(t_env *env, unsigned short *tab)
 {
 	int i;
 
@@ -66,9 +66,9 @@ static int blank_line(t_env *env, int *tab)
 int is_solvable(t_env *env)
 {
 	int end_inversions;
-	int *end_tab;
+	unsigned short *end_tab;
 	int start_inversions;
-	int *start_tab;
+	unsigned short *start_tab;
 
 	if (!(start_tab = malloc(sizeof(*start_tab) * (env->size * env->size))))
 	{
